# docs/admin/admin_swagger.yaml
definitions:
...
  dto.HsUserIdentityDeleteReq:
    properties:
      ids:
        items:
          type: integer
        type: array
    type: object
  dto.HsUserIdentityInsertReq:
    properties:
      createBy:
        type: integer
      credential:
        type: string
      identifier:
        type: string
      identityType:
        type: string
      updateBy:
        type: integer
      userId:
        type: string
      verified:
        type: string
    type: object
  dto.HsUserIdentityUpdateReq:
    properties:
      createBy:
        type: integer
      credential:
        type: string
      id:
        type: integer
      identifier:
        type: string
      identityType:
        type: string
      updateBy:
        type: integer
      userId:
        type: string
      verified:
        type: string
    type: object
...
  models.HsUserIdentity:
    properties:
      createBy:
        type: integer
      createdAt:
        type: string
      credential:
        type: string
      id:
        type: integer
      identifier:
        type: string
      identityType:
        type: string
      updateBy:
        type: integer
      updatedAt:
        type: string
      userId:
        type: string
      verified:
        type: string
    type: object
...
  response.Page:
    properties:
      count:
        type: integer
      pageIndex:
        type: integer
      pageSize:
        type: integer
    type: object
  response.Response:
    properties:
      code:
        type: integer
      msg:
        type: string
      requestId:
        description: 数据集
        type: string
      status:
        type: string
    type: object
...
info:
  contact: {}
  description: |-
    基于Gin + Vue + Element UI的前后端分离权限管理系统的接口文档
    添加qq群: 521386980 进入技术交流群 请先star，谢谢！
  license:
    name: MIT
    url: https://github.com/go-admin-team/go-admin/blob/master/LICENSE.md
  title: go-admin API
  version: 2.0.0
paths:
...
  /api/v1/hs-user-identity:
    delete:
      description: 删除HsUserIdentity
      parameters:
      - description: body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.HsUserIdentityDeleteReq'
      responses:
        "200":
          description: '{"code": 200, "message": "删除成功"}'
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: 删除HsUserIdentity
      tags:
      - HsUserIdentity
    get:
      description: 获取HsUserIdentity列表
      parameters:
      - description: 用户ID
        in: query
        name: userId
        type: string
      - description: '凭证类型: email, phone, google, apple, twitter, tiktok'
        in: query
        name: identityType
        type: string
      - description: '凭证值: 邮箱, 手机号, 第三方平台的唯一ID'
        in: query
        name: identifier
        type: string
      - description: 存储密码哈希(email/phone)或refresh_token(social)
        in: query
        name: credential
        type: string
      - description: 页条数
        in: query
        name: pageSize
        type: integer
      - description: 页码
        in: query
        name: pageIndex
        type: integer
      responses:
        "200":
          description: '{"code": 200, "data": [...]}'
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.Page'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/models.HsUserIdentity'
                        type: array
                    type: object
              type: object
      security:
      - Bearer: []
      summary: 获取HsUserIdentity列表
      tags:
      - HsUserIdentity
    post:
      consumes:
      - application/json
      description: 创建HsUserIdentity
      parameters:
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.HsUserIdentityInsertReq'
      responses:
        "200":
          description: '{"code": 200, "message": "添加成功"}'
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: 创建HsUserIdentity
      tags:
      - HsUserIdentity
  /api/v1/hs-user-identity/{id}:
    get:
      description: 获取HsUserIdentity
      parameters:
      - description: id
        in: path
        name: id
        type: integer
      responses:
        "200":
          description: '{"code": 200, "data": [...]}'
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.HsUserIdentity'
              type: object
      security:
      - Bearer: []
      summary: 获取HsUserIdentity
      tags:
      - HsUserIdentity
    put:
      consumes:
      - application/json
      description: 修改HsUserIdentity
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      - description: body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.HsUserIdentityUpdateReq'
      responses:
        "200":
          description: '{"code": 200, "message": "修改成功"}'
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: 修改HsUserIdentity
      tags:
      - HsUserIdentity
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"